+++ src/Phan/Language/Type/LiteralFloatType.php
@@ -2,8 +2,8 @@
 
 namespace Phan\Language\Type;
 
-use Phan\Config;
 use Phan\CodeBase;
+use Phan\Config;
 use Phan\Language\Context;
 use Phan\Language\Type;
 use RuntimeException;