+++ tests/files/src/0841_assign_op_creates_variable.php
@@ -1,0 +1,5 @@
+<?php
+function test() : float {
+    $x .= 'other';
+    return $x;
+}