+++ NEWS.md
@@ -8,6 +8,7 @@ New Features(Analysis):
 
 Bug fixes:
 + Fix a crash analyzing assignment operations on `$GLOBALS` such as `$GLOBALS['var'] += expr;` (#3615)
++ Fix false positive `Phan[Possibly]UndeclaredGlobalVariable` after conditions such as `assert($var instanceof MyClass` when the variable was not assigned to within the file or previously analyzed files. (#3616)
 
 Dec 29 2019, Phan 2.4.6
 -----------------------